+++ nofos/composer/views.py
@@ -1,7 +1,6 @@
 import json
 from typing import Dict
 
-from bloom_nofos import settings
 from bloom_nofos.html_diff import has_diff, html_diff
 from bloom_nofos.logs import log_exception
 from bloom_nofos.utils import generate_docx_download_response
@@ -17,11 +16,9 @@ from django.db import transaction
 from django.forms.models import model_to_dict
 from django.http import (
     Http404,
-    HttpResponse,
     HttpResponseBadRequest,
     HttpResponseForbidden,
     HttpResponseNotFound,
-    HttpResponseServerError,
 )
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse, reverse_lazy
@@ -38,7 +35,6 @@ from django.views.generic import (
     UpdateView,
     View,
 )
-from GrabzIt import GrabzItClient, GrabzItDOCXOptions
 from martor.utils import markdownify
 
 from nofos.audits import get_audit_event_by_id, safe_get_changed_fields
@@ -1808,50 +1804,19 @@ class WriterInstancePreviewView(BaseComposerPreviewView):
             return redirect(self.exit_url)
 
         if action == "download":
-            return self._handle_download_request(document, request)
+            export_url = request.build_absolute_uri(
+                reverse("composer:writer_instance_export", args=[document.pk])
+            )
+
+            return generate_docx_download_response(
+                request=request,
+                export_url=export_url,
+                target_element="#download_target",
+                filename_base=document.short_name or document.title,
+                tmp_name=str(document.pk),
+            )
 
         return HttpResponseBadRequest("Unknown action.")
-
-    def _handle_download_request(self, document, request):
-        """
-        Handle the download action for a ContentGuideInstance.
-        Makes a call to the Grabzit conversion API to convert the HTML to a DOCX file.
-        """
-        session_value = request.COOKIES["sessionid"]
-        csrf_value = request.COOKIES["csrftoken"]
-
-        grabzit = GrabzItClient.GrabzItClient(
-            settings.GRABZIT_APPLICATION_KEY, settings.GRABZIT_APPLICATION_SECRET
-        )
-
-        domain = request.get_host().split(":")[0]  # Remove port if present
-        set_session_cookie = grabzit.SetCookie("sessionid", domain, session_value)
-        set_csrf_cookie = grabzit.SetCookie("csrftoken", domain, csrf_value)
-
-        if not set_session_cookie or not set_csrf_cookie:
-            return HttpResponseServerError(
-                "Failed to set cookies for Grabzit conversion."
-            )
-
-        export_url = request.build_absolute_uri(
-            reverse("composer:writer_instance_export", args=[document.pk])
-        )
-        options = GrabzItDOCXOptions.GrabzItDOCXOptions()
-        options.targetElement = "#download_target"
-        grabzit.URLToDOCX(export_url, options)
-
-        filePath = f"/tmp/{document.pk}.docx"
-        grabzit.SaveTo(filePath)
-
-        with open(filePath, "rb") as docx_file:
-            response = HttpResponse(
-                docx_file.read(),
-                content_type="application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-            )
-            response["Content-Disposition"] = (
-                f'attachment; filename="{document.short_name or document.title}.docx"'
-            )
-            return response
 
 
 @method_decorator(staff_member_required, name="dispatch")